requirements.txt
# python>=3.10.0
--extra-index-url https://download.pytorch.org/whl/cu113
torch==1.12.1
--extra-index-url https://download.pytorch.org/whl/cu113
torchvision==0.13.1
tensorboard>=2.9.1
pillow>=9.0.1
tqdm>=4.63.0
scikit-learn>=1.2.2
seaborn>=0.12.2
